+++ src/app/books/page.tsx
@@ -1,0 +1,10 @@
+
+const Books = () => {
+    return (
+        <div>
+            Books
+        </div>
+    );
+};
+
+export default Books;