+++ src/components/index.tsx
@@ -1,0 +1,7 @@
+export {default as Header} from "./layout/Header";
+export {default as Footer} from "./layout/Footer";
+export {default as Content} from "./layout/Content";
+export {default as DarkMode} from "./layout/DarkMode";
+
+export {default as Container} from "./common/Container";
+export {default as ActiveLink} from "./common/ActiveLink";