+++ src/components/layout/Content.tsx
@@ -1,0 +1,15 @@
+import {FC, ReactNode} from 'react';
+
+interface IContentProps {
+    children: ReactNode;
+}
+
+const Content:FC<IContentProps> = ({children}) => {
+    return (
+        <main>
+            {children}
+        </main>
+    );
+};
+
+export default Content;