+++ CHANGE-LOG.md
@@ -4,8 +4,15 @@
 
 | Version            |
 |:-------------------|
+| [`0.2.0`](#v0.2.0) |
 | [`0.1.0`](#v0.1.0) |
 
+
+### <a id="v0.2.0"/>`0.2.0`
+
+- Added `react-icons`, `axios`, `sass`, and [`.nvmrc`](./.nvmrc) to the project
+- Completed the **header** and **navbar** section
+- Added [`DarkMode`](./src/components/layout/DarkMode.tsx) to handle the dark mode
 
 ### <a id="v0.1.0"/>`0.1.0`